Slience warning

+++ apple/objc/interface/database/WCTDatabase+Config.h
@@ -26,7 +26,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef NS_ENUM(NSInteger, WCTConfigPriority) {
+typedef NS_ENUM(int, WCTConfigPriority) {
     WCTConfigPriorityHigh = -100,
     WCTConfigPriorityDefault = 0,
     WCTConfigPriorityLow = 100,
